+++ src/components/access-type.js
@@ -395,7 +395,7 @@ class AccessTypeBase extends React.Component {
     if (!omise) {
       return Promise.reject({ message: "Payment option is loading..." });
     }
-    return omise.proceed(paymentObject).then((response) => response.proceed(paymentObject));
+    return omise.proceed(paymentObject).then((response) => response);
   };
 
   initAdyenPayment = (selectedPlanObj = {}, planType = "", AdyenModal, locale) => {